Cleaning up code and project; README updated and new iOS7 screenshots; fixing some deprecated warnings; other small fixes

+++ Demo/PhotoBrowserDemo/Menu.m
@@ -126,7 +126,7 @@
 }
 
 - (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
-    int rows = 0;
+    NSInteger rows = 0;
     
     if(section == 0)
         rows = 1;
